# Conflicts:
#	hazelcast-sql/src/test/java/com/hazelcast/jet/sql/impl/connector/kafka/model/JavaDeserializer.java
#	hazelcast-sql/src/test/java/com/hazelcast/jet/sql/impl/connector/kafka/model/JavaSerializer.java

+++ hazelcast/src/main/java/com/hazelcast/spi/impl/merge/AbstractContainerCollector.java
@@ -170,7 +170,7 @@ public abstract class AbstractContainerCollector<C> {
      *
      * @return the number of collected merge values
      */
-    protected abstract int getMergingValueCount();
+    protected abstract long getMergingValueCount();
 
     /**
      * Determines if the container should be merged.
